requirements.txt
highrise-bot-sdk>=23.1.0
python-dotenv>=1.0.1
discord.py>=2.4.0
